real time api spec renamed
